simplified code by removing CategoryDto.

+++ src/main/java/com/example/jobsearch/service/ResumeService.java
@@ -10,10 +10,14 @@ import com.example.jobsearch.model.Resume;
 import com.example.jobsearch.model.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
+
 @Slf4j
 @Service
 @RequiredArgsConstructor
@@ -23,15 +27,14 @@ public class ResumeService {
     private final WorkExperienceService workExperienceService;
     private final ContactInfoService contactInfoService;
     private final ApplicantProfileService applicantService;
-    private final CategoryService categoryService;
     private final UserService userService;
 
     private ResumeDto makeDtoFromResume(Resume resume) {
         return ResumeDto.builder()
                 .id(resume.getId())
-                .applicantDto(applicantService.getApplicantById(resume.getApplicantId()))
+                .authorEmail(applicantService.getApplicantById(resume.getApplicantId()).getUserId())
                 .resumeTitle(resume.getResumeTitle())
-                .categoryDto(categoryService.getCategoryByName(resume.getCategory()))
+                .category(resume.getCategory())
                 .expectedSalary(resume.getExpectedSalary())
                 .isActive(resume.isActive())
                 .isPublished(resume.isPublished())
@@ -41,9 +44,9 @@ public class ResumeService {
     private Resume createResumeFromDto(ResumeDto resume) {
         return Resume.builder()
                 .id(resume.getId())
-                .applicantId(resume.getApplicantDto().getId())
+                .applicantId(applicantService.findApplicantByUserId(resume.getAuthorEmail()).getId())
                 .resumeTitle(resume.getResumeTitle())
-                .category(resume.getCategoryDto().getCategory())
+                .category(resume.getCategory())
                 .expectedSalary(resume.getExpectedSalary())
                 .isActive(resume.isActive())
                 .isPublished(resume.isPublished())
@@ -79,22 +82,69 @@ public class ResumeService {
                 .toList();
     }
 
-    public void createResume(ResumeDto e, Authentication auth) {
+    public ResponseEntity<?> createResume(ResumeDto e, Authentication auth) {
         var u = auth.getPrincipal();
         User user = userService.getUserFromAuth(u.toString());
-        if(user.getId().equalsIgnoreCase(e.getApplicantDto().getUserId())) {
-            log.warn("New resume created: {}", e.getId());
-            resumeDao.save(createResumeFromDto(e));
+        if(user.getId().equalsIgnoreCase(e.getAuthorEmail())) {
+            Optional<Resume> r;
+            if(e.getId() == null) {
+            long x = (long)resumeDao.getAllResumes().size()+1;
+            r = resumeDao.getResumeById(x);
+            } else {
+                r = resumeDao.getResumeById(e.getId());
+            }
+            if(r.isEmpty()) {
+                resumeDao.save(createResumeFromDto(e));
+                return new ResponseEntity<>("Resume created successfully", HttpStatus.OK);
+            } else {
+                log.info("Tried to create a resume that already exists: {}", e.getId());
+                return new ResponseEntity<>("Resume already exists", HttpStatus.OK);
+            }
+        } else {
+            log.warn("Tried to create a resume for another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to create a resume for another user", HttpStatus.BAD_REQUEST);
+        }
+    }
+    private ResponseEntity<?> handleResumeQueries(Optional<Resume> maybeResume) {
+        if(maybeResume.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        return new ResponseEntity<>(makeDtoFromResume(maybeResume.get()), HttpStatus.OK);
+    }
+    public ResponseEntity<?> deleteResume(ResumeDto e, Authentication auth) {
+        var u = auth.getPrincipal();
+        User user = userService.getUserFromAuth(u.toString());
+        if(user.getId().equalsIgnoreCase(e.getAuthorEmail())) {
+            var r = resumeDao.getResumeById(e.getId());
+            if (r != null) {
+                resumeDao.delete(e.getId());
+                return new ResponseEntity<>("Resume deleted successfully", HttpStatus.OK);
+            } else {
+                log.info("Tried to delete a resume that does not exist: {}", e.getId());
+                return new ResponseEntity<>("Tried to delete a resume that does not exist", HttpStatus.OK);
+            }
+        } else {
+            log.warn("Tried to delete a resume of another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to delete a resume of another user", HttpStatus.BAD_REQUEST);
         }
     }
 
-    public void deleteResume(ResumeDto e) {
-        log.warn("Resume deleted: {}", e.getId());
-        resumeDao.delete(e.getId());
-    }
-
-    public void editResume(ResumeDto e) {
-        resumeDao.editResume(createResumeFromDto(e));
+    public ResponseEntity<?> editResume(ResumeDto e, Authentication auth) {
+        var u = auth.getPrincipal();
+        User user = userService.getUserFromAuth(u.toString());
+        if(user.getId().equalsIgnoreCase(e.getAuthorEmail())) {
+            var r = resumeDao.getResumeById(e.getId());
+            if (r != null) {
+                resumeDao.editResume(createResumeFromDto(e));
+            return new ResponseEntity<>("Resume edited successfully", HttpStatus.OK);
+        } else {
+                log.warn("Tried to edit a resume that does not exist: {}", e.getId());
+                return new ResponseEntity<>("Cannot edit a resume that does not exist", HttpStatus.NOT_FOUND);
+            }
+        } else {
+            log.warn("Tried to edit a resume of another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to edit a resume of another user", HttpStatus.BAD_REQUEST);
+        }
     }
 
     public List<ResumeDto> getAllResumesByLongList(List<Long> list) {
@@ -105,7 +155,7 @@ public class ResumeService {
     }
 
     public ResumeDto getResumeById(long id) {
-        return makeDtoFromResume(resumeDao.getResumeById(id));
+        return makeDtoFromResume(resumeDao.findResumeById(id));
     }
 
     public List<ResumeDto> getResumeByResumeTitle(String title) {
@@ -117,6 +167,7 @@ public class ResumeService {
 
     public void addEducation(EducationDto educationDto) {
         educationService.createEducation(educationDto);
+
     }
 
     public void editEducation(EducationDto educationDto) {
